+++ lib/date.dart
@@ -1,4 +1,4 @@
-// Copyright (c) 2015, <your name>. All rights reserved. Use of this source code
+// Copyright (c) 2015, <Tony Humbert>. All rights reserved. Use of this source code
 // is governed by a BSD-style license that can be found in the LICENSE file.
 
 /// The date library.
@@ -14,6 +14,7 @@ export 'src/myy.dart';
 export 'src/quarter.dart';
 export 'src/term.dart';
 export 'src/term_parse.dart';
+export 'src/time.dart';
 export 'src/time_iterable.dart';
 export 'src/utils.dart';
 export 'src/week.dart';