+++ test/all_test.dart
@@ -9,10 +9,12 @@ import 'interval_test.dart' as interval;
 import 'month_test.dart' as month;
 import 'term_test.dart' as term;
 import 'term_parse_test.dart' as term_parse;
+import 'time_test.dart' as time;
 import 'month_parse_test.dart' as month_parse;
 import 'quarter_test.dart' as quarter;
 import 'utils_test.dart' as utils;
 import 'week_test.dart' as week;
+
 
 Future<void> main() async {
   initializeTimeZones();
@@ -27,6 +29,7 @@ Future<void> main() async {
   quarter.tests();
   term.tests();
   term_parse.parseTermTest();
+  time.tests(); 
   utils.tests();
   week.tests();
 }